Item 5928 : Refactor ontology checks
- Bypass coherence checks (archive unit profile reference, type
compatibility & internal field delete) in forced update mode
- Refactoring

+++ sources/common/common-public/src/main/java/fr/gouv/vitam/common/model/administration/preservation/DefaultGriffin.java
@@ -50,7 +50,7 @@ public class DefaultGriffin {
 
     @Positive(message = ConstraintConstants.POSITIVE_ERROR_MSG)
     @JsonProperty("MaxSize")
-    private int maxSize;
+    private long maxSize;
 
     @NotNull(message = ConstraintConstants.NOT_NULL_ERROR_MSG)
     @JsonProperty("Debug")
@@ -88,11 +88,11 @@ public class DefaultGriffin {
         this.timeOut = timeOut;
     }
 
-    public int getMaxSize() {
+    public long getMaxSize() {
         return maxSize;
     }
 
-    public void setMaxSize(int maxSize) {
+    public void setMaxSize(long maxSize) {
         this.maxSize = maxSize;
     }
